Log cleanup: demote some warnings to debugs
It's expected behavior to not receive a response to the power query when
the device is off; suppress missing-ack warning in that scenario.

It's reported that newer TVs are hitting the "junk" warning log on
status query despite everything otherwise working as expected. Demote
that to a debug log too since there seems to be no ill effect.